pyproject.toml
[project]
name = 'wledcast'
version = '0.1.2'
description = 'Cast a screen to WLED'
authors = [{name = "Paul Pamment", email = "p.pamment@gmail.com"}]
license = {file = "LICENSE"}
readme = "README.md"

dependencies = [
    "asciimatics>=1.15,<2.0",
    "keyboard>=0.13.5,<1.0",
    "mss>=9.0",
    "numpy>=1.26,<2.0",
    "opencv-python>=4.8,<5.0",
    "pillow>=10.0",
    "pywinctl>=0.3,<1.0",
    "requests>=2.31.0",
    "wxpython>=4.2,<5.0",
    "zeroconf>=0.127.0",
]

[project.optional-dependencies]
dxcam = ["dxcam>=0.0.5"]

[metadata]
author = "Paul Pamment"
author_email = "p.pamment@gmail.com"
long_description = "file: README.md"
long_description_content_type = "text/markdown"
url = "https://github.com/ppamment/wledcast"
classifiers = """
    Programming Language :: Python :: 3
    License :: OSI Approved :: GNU General Public License v3 (GPLv3)
    Operating System :: Windows
    Operating System :: MacOS :: MacOS X
    Operating System :: POSIX :: Linux
"""

[tool.setuptools.package-data]
wledcast = ["filter.json"]

[project.scripts]
wledcast = 'wledcast.__main__:main'

[options]
python_requires = ">=3.10"
packages = "find: ."

[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

